requirements.txt
altair==5.2.0
attrs==23.2.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.11.0
decorator==5.1.1
dnspython==2.4.2
fonttools==4.51.0
idna==3.6
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.4
narwhals==1.47.0
nest-asyncio==1.6.0
numpy==1.24.4                   
packaging==23.2
pandas==2.1.4
pillow==10.2.0
platformdirs==3.11.0
plotly==5.21.0
protobuf==4.25.3
psutil==5.9.8
pyarrow==14.0.2
pydeck==0.8.1b0
pymongo==4.6.3
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
referencing==0.34.0
requests==2.31.0
rpds-py==0.10.6
scikit-learn==1.3.2
scipy==1.10.1                    
seaborn==0.12.2
six==1.16.0
streamlit==1.31.1
tenacity==8.2.3
threadpoolctl==3.4.0
toml==0.10.2
tornado==6.3.3
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
watchdog==3.0.0
xgboost==2.0.3
xlsxwriter==3.2.0